+++ AssetEditor.sln
@@ -50,6 +50,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ContentProject", "ContentPr
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "FBXWrapperNativeDll", "FBXWRapperDll\FBXWRapperDll.vcxproj", "{15AAE76D-0B83-45A2-B8AB-FD7BF38BDDC3}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AssetManagement", "AssetManagment\AssetManagement.csproj", "{413ABA15-75C4-4B2F-8B84-E5B8EADFE57F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -234,12 +236,23 @@ Global
 		{15AAE76D-0B83-45A2-B8AB-FD7BF38BDDC3}.Release|Any CPU.ActiveCfg = Release|x64
 		{15AAE76D-0B83-45A2-B8AB-FD7BF38BDDC3}.Release|x64.ActiveCfg = Release|x64
 		{15AAE76D-0B83-45A2-B8AB-FD7BF38BDDC3}.Release|x86.ActiveCfg = Release|Win32
+		{413ABA15-75C4-4B2F-8B84-E5B8EADFE57F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{413ABA15-75C4-4B2F-8B84-E5B8EADFE57F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{413ABA15-75C4-4B2F-8B84-E5B8EADFE57F}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{413ABA15-75C4-4B2F-8B84-E5B8EADFE57F}.Debug|x64.Build.0 = Debug|Any CPU
+		{413ABA15-75C4-4B2F-8B84-E5B8EADFE57F}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{413ABA15-75C4-4B2F-8B84-E5B8EADFE57F}.Debug|x86.Build.0 = Debug|Any CPU
+		{413ABA15-75C4-4B2F-8B84-E5B8EADFE57F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{413ABA15-75C4-4B2F-8B84-E5B8EADFE57F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{413ABA15-75C4-4B2F-8B84-E5B8EADFE57F}.Release|x64.ActiveCfg = Release|Any CPU
+		{413ABA15-75C4-4B2F-8B84-E5B8EADFE57F}.Release|x64.Build.0 = Release|Any CPU
+		{413ABA15-75C4-4B2F-8B84-E5B8EADFE57F}.Release|x86.ActiveCfg = Release|Any CPU
+		{413ABA15-75C4-4B2F-8B84-E5B8EADFE57F}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{FB2E6959-32F9-4F01-848B-21B38BDE98CC} = {18424B6A-CB8A-4CE1-935C-72459F31521B}
 		{98E09611-7C80-4413-9510-DE52EF71D7F1} = {18424B6A-CB8A-4CE1-935C-72459F31521B}
 		{3DC3A5F3-182F-46D4-BB7E-6639FC965E03} = {18424B6A-CB8A-4CE1-935C-72459F31521B}
 		{7FD31ACD-8D46-43FE-9E17-27198AED902B} = {C174C249-D109-49BE-8BC5-42A0D5DEDBDA}
@@ -250,6 +263,8 @@ Global
 		{3A205F23-1862-4F06-B9D8-DF82810515B3} = {FF70C4FA-A9A7-4DEC-87A3-4BF5F23EF336}
 		{F002EA23-09A0-42AE-9C5E-ADA6C1E95CC5} = {FF70C4FA-A9A7-4DEC-87A3-4BF5F23EF336}
 		{07D79931-CE1A-4092-AA6F-634282D31DC8} = {C174C249-D109-49BE-8BC5-42A0D5DEDBDA}
+		{15AAE76D-0B83-45A2-B8AB-FD7BF38BDDC3} = {C174C249-D109-49BE-8BC5-42A0D5DEDBDA}
+		{413ABA15-75C4-4B2F-8B84-E5B8EADFE57F} = {07AC615B-A8FC-4E1A-BDD5-BC11452429A0}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {AB5968F3-98ED-4AFF-98EA-0DBEDCACADF2}
