+++ AssetManagment/GenericFormats/SceneContainer.cs
@@ -1,28 +1,23 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using Matrix = Microsoft.Xna.Framework.Matrix;
 using Vector3 = Microsoft.Xna.Framework.Vector3;
 using Quaternion = Microsoft.Xna.Framework.Quaternion;
 
-namespace CommonControls.ModelFiles.FBX
+namespace AssetManagement.GenericFormats
 {
     public class SceneContainer
     {
-        public List<Mesh.PackedMesh> Meshes { get; set; } = new List<Mesh.PackedMesh>();
+        public List<PackedMesh> Meshes { get; set; } = new List<PackedMesh>();
         public List<BoneInfo> Bones { get; set; } = new List<BoneInfo>();
         public List<AnimationClip> Animations { get; set; } = new List<AnimationClip>();
-        public Node? RootNode { get; set; }
+        public Node RootNode { get; set; }
     }
 
-    
     public class Node
     {
-        public string Mame { get; set; }
-
+        public string Name { get; set; }
         public List<Node> Children { get; set; }
-
         public Node Parent { get; set; }
-
         public Matrix Transform { get; set; }
     }
 
@@ -31,16 +26,16 @@ namespace CommonControls.ModelFiles.FBX
         public string Name { get; set; }
         public int Id { get; set; }
         public int ParentId { get; set; }
-        Quaternion LocalRotation { get; set; }
-        Vector3 localTranslations { get; set; }
-        Matrix InverseBindPoseMatrix { get; set; }
+        public Quaternion LocalRotation { get; set; }
+        public Vector3 localTranslations { get; set; }
+        public Matrix InverseBindPoseMatrix { get; set; }
     }
 
     public struct AnimationKey
     {
         public Quaternion LocalRotation { get; set; }
         public Vector3 localTranslations { get; set; }
-        public Double TimeStamp { get; set; }
+        public double TimeStamp { get; set; }
     }
 
     public class NodeAnimation
@@ -63,6 +58,4 @@ namespace CommonControls.ModelFiles.FBX
             }
         }
     }
-
-
 };
