+++ CommonControls/Interfaces/AssetManagement/IAssetManagementFactory.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace CommonControls.Interfaces.AssetManagement
+{
+    public interface IAssetManagementFactory
+    {
+        List<IAssetImporter> GetAllImporters();
+        T GetImporter<T>() where T : IAssetImporter;
+        IAssetImporter GetImporter(string format);
+    }
+}
