+++ CommonControls/PackFileBrowser/OpenFileContexMenuHandler.cs
@@ -1,15 +1,11 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System.Collections.ObjectModel;
+﻿using System.Collections.ObjectModel;
 using CommonControls.Services;
 
 namespace CommonControls.PackFileBrowser
 {
     public class OpenFileContexMenuHandler : ContextMenuHandler
     {
-        public OpenFileContexMenuHandler(PackFileService service) : base(service, null)
+        public OpenFileContexMenuHandler(PackFileService service) : base(service, null, null)
         { }
 
         public override void Create(TreeNode node)
