+++ FBXWRapperDll/FBXWRapperDll.vcxproj.filters
@@ -976,7 +976,6 @@
     <None Include="Helpers\MS_SimpleMath\SimpleMath.inl">
       <Filter>Header Files</Filter>
     </None>
-    <None Include="ClassDiagram.cd" />
     <None Include="fbxsdk\bin\vs2019\x64\release\libfbxsdk.zip" />
     <None Include="fbxsdk\lib\vs2019\x64\debug\libfbxsdk.dll" />
     <None Include="fbxsdk\lib\vs2019\x64\release\libfbxsdk.dll" />
